+++ src/containerapp/azext_containerapp/_help.py
@@ -373,6 +373,11 @@ helps['containerapp add-on qdrant'] = """
     short-summary: Commands to manage the qdrant add-on for the Container Apps environment.
 """
 
+helps['containerapp add-on weaviate'] = """
+    type: group
+    short-summary: Commands to manage the weaviate add-on for the Container Apps environment.
+"""
+
 helps['containerapp add-on redis create'] = """
     type: command
     short-summary: Command to create the redis add-on.
@@ -398,6 +403,11 @@ helps['containerapp add-on qdrant create'] = """
     short-summary: Command to create the qdrant add-on.
 """
 
+helps['containerapp add-on weaviate create'] = """
+    type: command
+    short-summary: Command to create the weaviate add-on.
+"""
+
 helps['containerapp add-on redis delete'] = """
     type: command
     short-summary: Command to delete the redis add-on.
@@ -421,6 +431,11 @@ helps['containerapp add-on mariadb delete'] = """
 helps['containerapp add-on qdrant delete'] = """
     type: command
     short-summary: Command to delete the qdrant add-on.
+"""
+
+helps['containerapp add-on weaviate delete'] = """
+    type: command
+    short-summary: Command to delete the weaviate service.
 """
 
 helps['containerapp env update'] = """